Move scan page code to components/pages/scan
